+++ 1/package.json
@@ -9,21 +9,22 @@
 		"test": "vue-cli-service test:unit"
 	},
 	"dependencies": {
-		"apollo-server": "^2.9.12",
-		"apollo-server-testing": "^2.9.12",
+		"apollo-server": "^2.6.4",
+		"apollo-server-express": "^2.6.4",
+		"apollo-server-testing": "^2.6.4",
 		"canvas": "^2.5.0",
 		"core-js": "^3.1.2",
-		"graphql": "^14.5.8",
+		"graphql": "^14.3.1",
 		"vue": "^2.6.10",
 		"vue-mc": "^0.6.0"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.6.4",
-		"@vue/cli-plugin-babel": "^4.0.0",
-		"@vue/cli-plugin-eslint": "^4.0.0",
-		"@vue/cli-plugin-unit-jest": "^4.0.5",
-		"@vue/cli-service": "^4.0.0",
-		"@vue/test-utils": "^1.0.0-beta.29",
+		"@babel/core": "^7.7.4",
+		"@vue/cli-plugin-babel": "^4.1.1",
+		"@vue/cli-plugin-eslint": "^4.1.1",
+		"@vue/cli-plugin-unit-jest": "^4.1.1",
+		"@vue/cli-service": "^4.1.1",
+		"@vue/test-utils": "^1.0.0-beta.30",
 		"babel-eslint": "^10.0.1",
 		"babel-jest": "^24.9.0",
 		"eslint": "^5.16.0",
